Extend osName and osVersion to work on various systems
Fix #4
